+++ design/.storybook/preview.js
@@ -1,4 +1,4 @@
-import "../app.css";
+import "../styles.css";
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },